+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.2] - 2026-08-18
+
+### Added
+
+- Added an option to disable `obtain` unlocks entirely.
+- Added the ability to define category-level unlock criteria.
+
 ## [1.15.1] - 2026-08-14
 
 ### Fixed